backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.9.0
anthropic==0.34.2
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
reportlab==4.4.10
python-pptx==1.0.2
